Starting MVVM refactor with AddItem and Item Model.
