+++ app/src/main/java/com/example/ezmeal/Model/ItemModel.java
@@ -1,0 +1,25 @@
+package com.example.ezmeal.Model;
+
+//Data related to items
+
+public class ItemModel {
+    private String itemName;
+    private String brandName;
+
+    public String getItemName() {
+        return itemName;
+    }
+
+    public void setItemName(String itemName) {
+        this.itemName = itemName;
+    }
+
+    public String getBrandName() {
+        return brandName;
+    }
+
+    public void setBrandName(String brandName) {
+        this.brandName = brandName;
+    }
+
+}
