+++ app/src/main/java/com/example/ezmeal/Repository/AddListItemRepository.java
@@ -1,0 +1,7 @@
+package com.example.ezmeal.Repository;
+
+// A repository is a class where we fetch data from API or DB.
+
+public class AddListItemRepository {
+
+}
