+++ app/src/main/res/layout/fragment_add_list_item.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="16dp"
-    tools:context="navigationFragments.addItemView.AddListItemFragment">
+    tools:context=".View.AddListItemFragment">
 
     <!-- TODO: Update blank fragment layout -->
